pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mgl7760-projet-equipe3"
version = "0.1.0"
description = "Automatisation du projet Anfis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudpickle==3.1.2",
    "contourpy==1.3.3",
    "coverage==7.13.4",
    "cuda-bindings==12.9.4 ; sys_platform != 'darwin'",
    "cuda-pathfinder==1.4.2 ; sys_platform != 'darwin'",
    "cycler==0.12.1",
    "et-xmlfile==2.0.0",
    "filelock==3.25.2",
    "fonttools==4.62.1",
    "fsspec==2026.2.0",
    "imbalanced-learn==0.14.1",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "jupyter==1.1.1",
    "jupyterlab==4.4.9",
    "kiwisolver==1.5.0",
    "notebook==7.4.7",
    "llvmlite==0.46.0 ; platform_machine != 'x86_64' or sys_platform != 'darwin'",
    "llvmlite>=0.45.0,<0.46 ; platform_machine == 'x86_64' and sys_platform == 'darwin'",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "mlflow>=3.0.0,<4.0",
    "mpmath==1.3.0",
    "networkx==3.6.1",
    "numba==0.64.0 ; platform_machine != 'x86_64' or sys_platform != 'darwin'",
    "numba>=0.62.0,<0.63 ; platform_machine == 'x86_64' and sys_platform == 'darwin'",
    "numpy==2.4.3 ; platform_machine != 'x86_64' or sys_platform != 'darwin'",
    "numpy>=2.0,<2.4 ; platform_machine == 'x86_64' and sys_platform == 'darwin'",
    "nvidia-cublas-cu12==12.8.4.1 ; sys_platform != 'darwin'",
    "nvidia-cuda-cupti-cu12==12.8.90 ; sys_platform != 'darwin'",
    "nvidia-cuda-nvrtc-cu12==12.8.93 ; sys_platform != 'darwin'",
    "nvidia-cuda-runtime-cu12==12.8.90 ; sys_platform != 'darwin'",
    "nvidia-cudnn-cu12==9.10.2.21 ; sys_platform != 'darwin'",
    "nvidia-cufft-cu12==11.3.3.83 ; sys_platform != 'darwin'",
    "nvidia-cufile-cu12==1.13.1.3 ; sys_platform != 'darwin'",
    "nvidia-curand-cu12==10.3.9.90 ; sys_platform != 'darwin'",
    "nvidia-cusolver-cu12==11.7.3.90 ; sys_platform != 'darwin'",
    "nvidia-cusparse-cu12==12.5.8.93 ; sys_platform != 'darwin'",
    "nvidia-cusparselt-cu12==0.7.1 ; sys_platform != 'darwin'",
    "nvidia-nccl-cu12==2.27.5 ; sys_platform != 'darwin'",
    "nvidia-nvjitlink-cu12==12.8.93 ; sys_platform != 'darwin'",
    "nvidia-nvshmem-cu12==3.4.5 ; sys_platform != 'darwin'",
    "nvidia-nvtx-cu12==12.8.90 ; sys_platform != 'darwin'",
    "openpyxl==3.1.5",
    "packaging==26.0",
    "pandas>=2.2,<3.0",
    "pillow==12.1.1",
    "pluggy==1.6.0",
    "pygments==2.19.1",
    "pyparsing==3.3.2",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "pytest-html>=4.1.1",
    "python-dateutil==2.9.0.post0",
    "scikit-fuzzy==0.5.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "seaborn==0.13.2",
    "setuptools==82.0.1",
    "shap==0.51.0",
    "six==1.17.0",
    "sklearn-compat==0.1.5",
    "slicer==0.0.8",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "torch==2.10.0",
    "tqdm==4.67.3",
    "triton==3.6.0 ; sys_platform != 'darwin'",
    "typing-extensions==4.15.0",
    "xlrd==2.0.2",
]

# Development dependencies for code quality tools
[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",        # Fast Python linter and formatter (replaces black)
    "mypy>=1.13.0",       # Static type checking
    "pre-commit>=3.0.0",  # Git hooks for automated checks
    "bandit[toml]>=1.7.0", # Security linter for Python
    "pip-audit>=2.0.0",   # Check for known vulnerabilities in dependencies
    "mutmut>=2.5.1",      # Mutation testing
    "sphinx>=8.0.0",      # API documentation (Sphinx + autodoc)
    "sphinx-rtd-theme>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]

# Ruff configuration - modern Python linter and formatter (replaces black + flake8)
[tool.ruff]
target-version = "py312"
line-length = 88  # Standard Python formatting width
src = ["src", "tests"]
exclude = [
    ".git",
    "__pycache__", 
    "docs",
    "build",
    "*.egg-info",
    ".venv",
]

[tool.ruff.lint]
# Select rules we want to enforce
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort imports
    "N",   # naming conventions
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
# Ignore some rules that might be too strict for learning / ML conventions
ignore = [
    "E501",  # Line too long (handled by ruff format)
    "N803",  # Argument name should be lowercase (allow X, y for ML)
    "N806",  # Variable name should be lowercase (allow X_train for ML)
]

# Notebooks: allow imports after comments/setup in same cell
[tool.ruff.lint.per-file-ignores]
"credit_risk_analysis.ipynb" = ["E402"]

# Ruff formatting configuration (replaces Black)
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy configuration - type checking
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
strict_optional = true
# Be lenient with missing imports for third-party packages
ignore_missing_imports = true

# Pytest: run only tests in tests/ (exclude src/test_*.py scripts that need dataset)
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "-q",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=html:reports/coverage_html",
    "--junitxml=reports/junit.xml",
    "--html=reports/pytest_report.html",
    "--self-contained-html",
    "--cov-fail-under=65",
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "src/main_pipeline.py",
    "src/credit_student_pipeline.py",
    "src/test_anfis.py",
    "src/test_pipeline.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
omit = [
    "src/main_pipeline.py",
    "src/credit_student_pipeline.py",
    "src/test_anfis.py",
    "src/test_pipeline.py",
]

# Bandit configuration - security linting
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
# Skip some checks that might be too restrictive for learning projects
skips = ["B101"]  # Skip assert_used - allows assert in tests
